fix: post SponsorBtn,list 수정중
