+++ src/components/post/SponsorBtn.jsx
@@ -1,16 +1,16 @@
-import React, { useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { auth, db } from '../../firebase';
 import styled from 'styled-components';
-import { addDoc, collection } from 'firebase/firestore';
+import { addDoc, collection, getDocs, query } from 'firebase/firestore';
 import SponsorTimeLine from 'components/SponsorTimeLine';
 import SponsorPercent from 'components/SponsorPercent';
 import HeartButton from './HeartButton';
 
-const SponsorBtn = () => {
+const SponsorBtn = ({ receiptPrice, setReceiptPrice }) => {
   const user = auth.currentUser;
 
   const [isAdd, setIsAdd] = useState(false);
-  const [receiptPrice, setReceiptPrice] = useState(0);
+  const [totalPrice, setTotalPrice] = useState(0);
 
   const onChangeReceipt = (event) => {
     const rawValue = event.target.value;
@@ -45,11 +45,12 @@ const SponsorBtn = () => {
       <Achieve>
         <div>
           <SponsorPercent />
-          <SponsorTimeLine />
+          {/* <PointText color="var(--sub-color)">{totalPrice}</PointText>원 모금 */}
           <PointText color="var(--main-color)">98%&nbsp;</PointText>달성
         </div>
         <div>
-          <PointText color="var(--sub-color)">123123&nbsp;</PointText>원 달성
+          {/* <PointText color="var(--sub-color)">123123&nbsp;</PointText>원 달성 */}
+          <SponsorTimeLine totalPrice={totalPrice} setTotalPrice={setTotalPrice} />
         </div>
       </Achieve>
       <PriceForm onSubmit={handleOnSubmit}>
