+++ src/components/post/SponsorList.jsx
@@ -1,47 +1,55 @@
-import React from 'react';
-import defaultUser from 'assets/defaultUser.png';
+import React, { useEffect } from 'react';
 import styled from 'styled-components';
+import SponsorItem from './SponsorItem';
+import { collection, getDocs, orderBy, query } from 'firebase/firestore';
+import { db } from '../../firebase';
 
-const SponsorList = ({ username, receiptPrice, profile }) => {
+const SponsorList = ({ userComment, setUserComment }) => {
+  useEffect(() => {
+    const fetchUserComment = async () => {
+      const userCommentQuery = query(collection(db, 'sponsorUser'), orderBy('createdAt', 'desc'));
+      const spanshot = await getDocs(userCommentQuery);
+      const userComment = spanshot.docs.map((doc) => {
+        const { profile, receiptPrice, username, createdAt } = doc.data();
+        return {
+          createdAt,
+          profile,
+          receiptPrice,
+          username
+        };
+      });
+      setUserComment(userComment);
+    };
+    fetchUserComment();
+  }, [setUserComment]);
   return (
     <>
-      <CommentWrapper>
-        <UserImg src={profile ?? defaultUser} alt="User Profile" />
-        <CommentText>
-          {username ?? '유저 닉네임'}님이 &nbsp; <FontWeight>{receiptPrice.toLocaleString('ko-KR')}원 펀딩</FontWeight>
-          해 주셨어요.
-        </CommentText>
-      </CommentWrapper>
+      {userComment.length === 0 ? (
+        <div>
+          <p>아직 참여 중인 서포터가 없습니다.</p>
+          <p>첫 번째 서포터가 되어보세요!</p>
+        </div>
+      ) : null}
+      <CommentContainer>
+        {userComment.map((item) => (
+          <SponsorItem key={item.id} {...item} />
+        ))}
+      </CommentContainer>
     </>
   );
 };
 
 export default SponsorList;
 
-const CommentWrapper = styled.div`
+const CommentContainer = styled.div`
   display: flex;
+  flex-direction: column;
+  justify-content: center;
   align-items: center;
-  border: 2px solid #dfdfdf;
-  width: 800px;
-  margin: 7px auto;
-  height: 70px;
-  border-radius: 30px;
-  background-color: white;
-`;
-
-const UserImg = styled.img`
-  margin-left: 30px;
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  margin-right: 15px;
-`;
-
-const CommentText = styled.p`
-  font-size: 16px;
-  display: flex;
-`;
-
-const FontWeight = styled.p`
+  margin: 50px auto;
   font-weight: bold;
+  line-height: 2;
+  font-size: 16px;
+  gap: 15px;
+  margin: 50px auto;
 `;
