+++ src/pages/Post.jsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import Navbar from 'components/common/Navbar';
 import styled from 'styled-components';
-import ProductsList from 'data/products.json';
 import SponsorBtn from 'components/post/SponsorBtn';
 import SponsorItem from 'components/post/SponsorItem';
 import ScheduledNotification from 'components/post/ScheduledNotification';
@@ -11,6 +10,7 @@ import CompletedComments from 'components/post/CompletedComments';
 import { useParams } from 'react-router';
 import { collection, getDocs, query, updateDoc, doc } from '@firebase/firestore';
 import { db } from '../firebase';
+import SponsorList from 'components/post/SponsorList';
 
 const ProjectIntroduction = styled.div`
   display: flex;
@@ -99,6 +99,8 @@ function Post({ activeNavTab, setActiveNavTab }) {
   const [activePostTab, setActivePostTab] = useState('project');
   const [projects, setProject] = useState([]);
   const id = useParams().id;
+  const [receiptPrice, setReceiptPrice] = useState(0);
+  const [userComment, setUserComment] = useState([]);
 
   // DB에서 데이터 가져오기
   useEffect(() => {
@@ -177,7 +179,7 @@ function Post({ activeNavTab, setActiveNavTab }) {
             onCancelOpenNotification={handleCancelOpenNotification}
           />
 
-          {/* <SponsorBtn /> */}
+          <SponsorBtn receiptPrice={receiptPrice} setReceiptPrice={setReceiptPrice} />
 
           {/* <CompletedNotification /> */}
         </TitleBox>
@@ -195,9 +197,9 @@ function Post({ activeNavTab, setActiveNavTab }) {
         {activePostTab === 'project' ? (
           <ProjectInfoContainer dangerouslySetInnerHTML={dangerousHTML} />
         ) : (
-          <ScheduledComments />
+          // <ScheduledComments />
 
-          // <SponsorItem />
+          <SponsorList userComment={userComment} setUserComment={setUserComment} />
 
           // <CompletedComments />
         )}
